Removed NewsSection and Split-out Risk and Ai Summary Sections
